try to make PrivateUsers work with MountAPIVFS=no
this changes the userns child process to be started *before* all
other namespaces, in particular the mount namespace, are set up.
this means that it is guaranteed to be able to write /proc, even
if the mount namespace does not include it anymore. partially
tested, both normal and integration tests seem to pass. i'm
committing this now before doing further testing.